:recycle: Refactor: pacakge convention에 따른 naming 변경
