+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/problemteam/ProblemTeam.java
@@ -1,6 +1,7 @@
-package com.unsolvedwa.unsolvedwa.domain.user_team;
+package com.unsolvedwa.unsolvedwa.domain.problemteam;
 
 import com.unsolvedwa.unsolvedwa.domain.BaseTimeEntity;
+import com.unsolvedwa.unsolvedwa.domain.problem.Problem;
 import com.unsolvedwa.unsolvedwa.domain.team.Team;
 import com.unsolvedwa.unsolvedwa.domain.user.User;
 import lombok.Getter;
@@ -9,12 +10,18 @@ import javax.persistence.*;
 
 @Entity
 @Getter
-public class UserTeam extends BaseTimeEntity {
+public class ProblemTeam extends BaseTimeEntity {
 
   @Id
   @GeneratedValue
-  @Column(name = "user_team_id")
+  @Column(name = "problem_team_id")
   private Long id;
+
+  private Long score;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "problem_id")
+  private Problem problem;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "team_id")
@@ -23,6 +30,4 @@ public class UserTeam extends BaseTimeEntity {
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id")
   private User user;
-
-  private Long score;
 }
