+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/team/Team.java
@@ -2,8 +2,8 @@ package com.unsolvedwa.unsolvedwa.domain.team;
 
 import com.unsolvedwa.unsolvedwa.domain.BaseTimeEntity;
 import com.unsolvedwa.unsolvedwa.domain.ranking.Ranking;
-import com.unsolvedwa.unsolvedwa.domain.problem_team.ProblemTeam;
-import com.unsolvedwa.unsolvedwa.domain.user_team.UserTeam;
+import com.unsolvedwa.unsolvedwa.domain.problemteam.ProblemTeam;
+import com.unsolvedwa.unsolvedwa.domain.userteam.UserTeam;
 import lombok.Getter;
 
 import javax.persistence.*;
