+++ src/main/java/com/unsolvedwa/unsolvedwa/domain/userteam/UserTeam.java
@@ -1,7 +1,6 @@
-package com.unsolvedwa.unsolvedwa.domain.problem_team;
+package com.unsolvedwa.unsolvedwa.domain.userteam;
 
 import com.unsolvedwa.unsolvedwa.domain.BaseTimeEntity;
-import com.unsolvedwa.unsolvedwa.domain.problem.Problem;
 import com.unsolvedwa.unsolvedwa.domain.team.Team;
 import com.unsolvedwa.unsolvedwa.domain.user.User;
 import lombok.Getter;
@@ -10,18 +9,12 @@ import javax.persistence.*;
 
 @Entity
 @Getter
-public class ProblemTeam extends BaseTimeEntity {
+public class UserTeam extends BaseTimeEntity {
 
   @Id
   @GeneratedValue
-  @Column(name = "problem_team_id")
+  @Column(name = "user_team_id")
   private Long id;
-
-  private Long score;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "problem_id")
-  private Problem problem;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "team_id")
@@ -30,4 +23,6 @@ public class ProblemTeam extends BaseTimeEntity {
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id")
   private User user;
+
+  private Long score;
 }
